Extract SimpleListItem
